+++ include/thrust/system/cuda/window_2d.inl
@@ -9,7 +9,7 @@ namespace thrust
   }
 
   template <class T,class Alloc>
-  __host__ __device__ window_2d<T,Alloc>::window_2d(block_2d<T,Alloc> *b,size_t start_x, size_t start_y, size_t window_dim_x, size_t window_dim_y)
+  __host__ __device__ window_2d<T,Alloc>::window_2d(block_2d<T,Alloc> *b,int start_x, int start_y, int window_dim_x, int window_dim_y)
   {
     // TODO: Better Boundary checks.
     this->start_x = start_x;
@@ -43,7 +43,7 @@ namespace thrust
     this->is_texture = obj.is_texture;
   }
   template <class T, class Alloc>
-  __host__ __device__ window_2d<T,Alloc>::window_2d(cudaTextureObject_t texref, size_t start_x, size_t start_y, size_t window_dim_x, size_t window_dim_y)
+  __host__ __device__ window_2d<T,Alloc>::window_2d(cudaTextureObject_t texref, int start_x, int start_y, int window_dim_x, int window_dim_y)
   {
 
     this->start_x = start_x;
@@ -56,7 +56,7 @@ namespace thrust
   }
 
   template <class T,class Alloc>
-  __host__ __device__ window_2d<T,Alloc>::window_2d(block_2d<T,Alloc> *b, T *data,size_t start_x, size_t start_y, size_t local_start_x, size_t local_start_y, size_t window_dim_x, size_t window_dim_y, size_t block_dim_x, size_t block_dim_y,size_t pitch)
+  __host__ __device__ window_2d<T,Alloc>::window_2d(block_2d<T,Alloc> *b, T *data,int start_x, int start_y, int local_start_x, int local_start_y, int window_dim_x, int window_dim_y, int block_dim_x, int block_dim_y,int pitch)
   {
     // TODO: Better Boundary checks.
     this->start_x = start_x;
@@ -78,7 +78,7 @@ namespace thrust
   }
 
   template <class T>
-  __host__ __device__ window_2d_iterator<T>::window_2d_iterator<T>(T * data, size_t position)
+  __host__ __device__ window_2d_iterator<T>::window_2d_iterator<T>(T * data, int position)
   {
     this->data = data;
     this->position = position;
@@ -87,9 +87,9 @@ namespace thrust
 
 
   template <class T,class Alloc>
-  __host__ __device__ window_2d_iterator<T> window_2d<T,Alloc>::operator[] (size_t index) const
+  __host__ __device__ window_2d_iterator<T> window_2d<T,Alloc>::operator[] (int index) const
   {
-    size_t position;
+    int position;
     if(this->is_shared)
     {
       position = (local_start_y+index)*block_dim_x + local_start_x;
@@ -120,13 +120,13 @@ namespace thrust
     }
   }
   template<class T>
-  __host__ __device__ T & window_2d_iterator<T>::operator[] (size_t index) const
+  __host__ __device__ T & window_2d_iterator<T>::operator[] (int index) const
   {
       return data[position + index];
   }
 
   template <class T,class Alloc>
-  __host__ window_iterator<T,Alloc>::window_iterator(block_2d<T,Alloc> *b, size_t window_dim_x, size_t window_dim_y, size_t stride_x, size_t stride_y)
+  __host__ window_iterator<T,Alloc>::window_iterator(block_2d<T,Alloc> *b, int window_dim_x, int window_dim_y, int stride_x, int stride_y)
   {
     this->b = b->device_pointer;
     this->data_pointer = b->data_pointer;
@@ -138,12 +138,12 @@ namespace thrust
     this->stride_y = stride_y;
     this->pitch = b->pitch;
     this->position =0;
-    this->windows_along_x= size_t((this->block_dim_x-window_dim_x)/stride_x) +1;
-    this->windows_along_y = size_t((this->block_dim_y-window_dim_y)/stride_y)+1;
+    this->windows_along_x= int((this->block_dim_x-window_dim_x)/stride_x) +1;
+    this->windows_along_y = int((this->block_dim_y-window_dim_y)/stride_y)+1;
   }
 
   template <class T,class Alloc>
-  __host__ window_iterator<T,Alloc>::window_iterator(block_2d<T,Alloc> *b, size_t window_dim_x, size_t window_dim_y, size_t stride_x, size_t stride_y,size_t position)
+  __host__ window_iterator<T,Alloc>::window_iterator(block_2d<T,Alloc> *b, int window_dim_x, int window_dim_y, int stride_x, int stride_y,int position)
   {
     this->b = b->device_pointer;
     this->data_pointer = b->data_pointer;
@@ -155,17 +155,17 @@ namespace thrust
     this->stride_x = stride_x;
     this->stride_y = stride_y;
     this->position =position;
-    this->windows_along_x= size_t((this->block_dim_x-window_dim_x)/stride_x) +1;
-    this->windows_along_y = size_t((this->block_dim_y-window_dim_y)/stride_y)+1;
+    this->windows_along_x= int((this->block_dim_x-window_dim_x)/stride_x) +1;
+    this->windows_along_y = int((this->block_dim_y-window_dim_y)/stride_y)+1;
   }
 
   template <class T,class Alloc>
   __host__ __device__ window_iterator<T,Alloc>::reference window_iterator<T,Alloc>::operator* () const
   {
-    size_t i = position%windows_along_x;
-    size_t j = position/windows_along_x;
-    size_t start_x = stride_x*i;
-    size_t start_y = stride_y*j;
+    int i = position%windows_along_x;
+    int j = position/windows_along_x;
+    int start_x = stride_x*i;
+    int start_y = stride_y*j;
     window_2d<T,Alloc> temp(b, start_x,start_y,this->window_dim_x, this->window_dim_y);
     return temp;
   }
@@ -286,7 +286,7 @@ namespace thrust
   }
 
   template <class T,class Alloc>
-  window_vector<T,Alloc>::window_vector (block_2d<T,Alloc> *b, size_t window_dim_x, size_t window_dim_y, size_t stride_x, size_t stride_y)
+  window_vector<T,Alloc>::window_vector (block_2d<T,Alloc> *b, int window_dim_x, int window_dim_y, int stride_x, int stride_y)
   {
     this->b = b;
     this->window_dim_x = window_dim_x;
@@ -305,8 +305,8 @@ namespace thrust
   window_iterator<T,Alloc> window_vector<T,Alloc>::end()
   {
 
-    size_t windowsX = size_t((b->dim_x-window_dim_x)/stride_x) +1;
-    size_t windowsY = size_t((b->dim_y-window_dim_y)/stride_y)+1;
+    int windowsX = int((b->dim_x-window_dim_x)/stride_x) +1;
+    int windowsY = int((b->dim_y-window_dim_y)/stride_y)+1;
     return window_iterator<T,Alloc>(b,window_dim_x,window_dim_y,stride_x,stride_y,windowsX*windowsY);
   }
